requirements.txt
youtube_dl
coverage
pysrt
guess_language-spirit
